backend/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astroid==2.6.5
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
autopep8==1.5.7
bcrypt==3.2.0
blessed==1.18.1
boto3==1.18.13
botocore==1.21.13
cached-property==1.5.2
cachetools==4.2.2
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
exponent-server-sdk==2.0.0
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.16.0
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.3
isort==5.9.3
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
paramiko==2.7.2
parso==0.7.1
pathspec==0.5.9
Pillow==8.3.1
pluggy==0.13.1
pooch==1.4.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
PyJWT==1.7.1
pylint==2.9.6
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rope==0.19.0
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
semantic-version==2.8.5
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==2.2.0
toml==0.10.2
twilio==6.62.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
zipp==3.5.0
